Fix specs, rename ui and add linting to codeship
