+++ ui/.eslintrc
@@ -3,6 +3,7 @@
   "extends": "airbnb",
   "rules": {
     "comma-dangle": [2, "never"],
-    "object-shorthand": [2, "never"]
+    "object-shorthand": [2, "never"],
+    "no-underscore-dangle": ["error", { "allowAfterThis": true }]
   }
 }
