+++ ui/specs/components/vote_spec.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import TestUtils from 'react/lib/ReactTestUtils';
 import Vote from './../../src/components/vote';
-import Ui from './../../src/views/ui/ui';
+import Ui from './../../src/views/common/ui';
 
 describe('Vote', () => {
   let instance;
