+++ ui/src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render } from 'react-dom';
 import { Router, Route, browserHistory } from 'react-router';
-import UI from './views/ui/ui';
+import UI from './views/common/ui';
 import Playlist from './views/subviews/playlist';
 import Account from './views/subviews/account';
