+++ ui/src/views/common/navigation.js
