+++ ui/src/views/common/side-panel.js
